Correção de Erro de Compilação em D7
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@24139 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrDFe/ACBrMDFe/PCNMDFe/pmdfeEnvEventoMDFe.pas
@@ -485,7 +485,7 @@ begin
   try
     LerIniArquivoOuString(AIniString, INIRec);
 
-    idLote := INIRec.ReadInt64('EVENTO', 'idLote', 0);
+    idLote := INIRec.ReadInteger('EVENTO', 'idLote', 0);
 
     I := 1;
     while true do
