+++ Fontes/ACBrDFe/ACBrNF3e/PCNNF3e/pcnEnvEventoNF3e.pas
@@ -361,7 +361,7 @@ begin
   INIRec := TMemIniFile.Create('');
   try
     LerIniArquivoOuString(AIniString, INIRec);
-    idLote := INIRec.ReadInt64('EVENTO', 'idLote', 0);
+    idLote := INIRec.ReadInteger('EVENTO', 'idLote', 0);
 
     I := 1 ;
     while true do
