+++ Fontes/ACBrDFe/ACBrNFe/PCNNFe/pcnEnvEventoNFe.pas
@@ -565,7 +565,7 @@ begin
   INIRec := TMemIniFile.Create('');
   try
     LerIniArquivoOuString(AIniString, INIRec);
-    idLote := INIRec.ReadInt64( 'EVENTO','idLote' ,INIRec.ReadInteger( 'CCE','idLote',0));
+    idLote := INIRec.ReadInteger( 'EVENTO','idLote' ,INIRec.ReadInteger( 'CCE','idLote',0));
 
     I := 1 ;
     while true do
